connect form onsubmit to buttons
